% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pane.r
\name{init.yaml.pane}
\alias{init.yaml.pane}
\title{Initialize a pane specified with yaml code}
\usage{
init.yaml.pane(yaml = NULL, pane = NULL)
}
\description{
Initialize a pane specified with yaml code
}

