% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lines_utils.r
\name{lines.intersections}
\alias{lines.intersections}
\title{Find the intersections of two curves, which are characterized by their xy values}
\usage{
\method{lines}{intersections}(line1, line2, grid.length = 201)
}
\description{
Find the intersections of two curves, which are characterized by their xy values
}

