% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoms.r
\name{objects.to.geoms}
\alias{objects.to.geoms}
\title{Convert a list of abstract geometrical objects to a list of geoms}
\usage{
objects.to.geoms(objs, params, xrange, yrange, ...)
}
\description{
Convert a list of abstract geometrical objects to a list of geoms
}

