% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.r
\name{plot.curve}
\alias{plot.curve}
\title{plot a single curve in a pane}
\usage{
\method{plot}{curve}(curve = NULL, eq = NULL, pane, main = NULL,
  label = "", ...)
}
\description{
plot a single curve in a pane
}

