% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pane.r
\name{plot.pane}
\alias{plot.pane}
\title{Plot a pane}
\usage{
\method{plot}{pane}(pane, geoms = pane$geoms, xrange = pane$xrange,
  yrange = pane$yrange, alpha = 1, main = "", mar = c(4, 3, 1, 1),
  show.grid = !TRUE, label.df = NULL, lwd.factor = 1, label.cex = 0.75,
  cex.axis = 0.8)
}
\description{
Plot a pane
}

