% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_utils.r
\name{two.point.lines.intersections}
\alias{two.point.lines.intersections}
\title{intersection of two geoms that are characterized by two points each}
\usage{
two.point.lines.intersections(x1, x2, x3, x4, y1, y2, y3, y4)
}
\description{
Formula based on Wikipedia entry
http://en.wikipedia.org/wiki/Line%E2%80%93geom_intersection
}

