% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pane.r
\name{init.pane}
\alias{init.pane}
\title{Initilize a pane}
\usage{
init.pane(name = NULL, xvar = NULL, yvar = NULL, xrange = NULL,
  yrange = NULL, xmarkers = NULL, ymarkers = NULL, lines = NULL,
  curves = NULL, pane = list(), init.curves = TRUE)
}
\description{
Initilize a pane
}

