% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.r
\name{object.to.geom}
\alias{object.to.geom}
\title{Convert an abstract geometrical object to a geom}
\usage{
object.to.geom(obj, values = pane$values, xrange = pane$xrange,
  yrange = pane$yrange, xlen = 201, ylen = 201, pane = NULL, ...)
}
\description{
Convert an abstract geometrical object to a geom
}

