% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoms.r
\name{object.to.geom}
\alias{object.to.geom}
\title{Convert an abstract geometrical object to a geom}
\usage{
object.to.geom(obj, params, xrange, yrange, ...)
}
\description{
Convert an abstract geometrical object to a geom
}

